refactor: add components folder in home page
